+++ libraries/matchers.rb
@@ -25,6 +25,7 @@ if defined?(ChefSpec)
     oneview_enclosure:                  [:add, :remove, :refresh, :reconfigure, :patch],
     oneview_enclosure_group:            standard_actions + [:set_script],
     oneview_ethernet_network:           standard_actions + [:reset_connection_template],
+    oneview_event:                      [:create],
     oneview_fabric:                     [:set_reserved_vlan_range],
     oneview_fc_network:                 standard_actions,
     oneview_fcoe_network:               standard_actions,